requirements.txt
pandas
requests
python-dotenv
streamlit
streamlit-lottie